# msatendimento/src/main/java/com/api/msatendimento/model/DTO/VacinaDTO.java
package com.api.msatendimento.model.DTO;


public class VacinaDTO {

    private Long id;
    private String nome;
    private String marca;

    public VacinaDTO() {
    }

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getNome() {
        return nome;
    }

    public void setNome(String nome) {
        this.nome = nome;
    }

    public String getMarca() {
        return marca;
    }

    public void setMarca(String marca) {
        this.marca = marca;
    }
}