# msvacina/src/main/resources/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "api de vacina"
  description: "Tem como objetivo salvar dados e listar vacinas"
host: "localhost:8080"
schemes:
  - https
tags:
  - name: "VacinaController"
paths:
  /vacina/{nome}:
    get:
      tags:
        - "VacinaController"
      summary: "buscar vacina por nome"
      operationId: "findByName"
      produces:
        - application/json
      description: "Esta operação tem como objetivo buscar vacina por nome"
      parameters:
        - in: "path"
          name: "nome"
          type: "string"
          required: true
          description: "Nome da Vacina"
      responses:
        200:
          description: "ok"
          schema:
            $ref: "#/definitions/VacinaDTO"
        400:
          description: "Consulta inválida"
        401:
          description: "Não foi possível credenciais para autenticação inválida"
        404:
          description: "Vacina não encontrada"
        500:
          description: "Erro interno"
  /vacina:
    post:
      tags:
        - "VacinaController"
      summary: "Cadastrar vacina"
      operationId: "save"
      produces:
        - application/json
      description: "Esta operacao tem como objetivo cadastrar vacina"
      parameters:
        - in: "body"
          name: "vacinaRequest"
          required: true
          description: "Cadastrar vacina"
          schema:
            $ref: "#/definitions/VacinaRequest"
      responses:
        201:
          description: "create"
        400:
          description: "Consulta inválida"
        401:
          description: "Não foi possível credenciais para autenticação inválida"
        404:
          description: "Vacina não encontrada"
        500:
          description: "Erro interno"

  /vacina/delete/{id}:
    delete:
      tags:
        - "VacinaController"
      operationId: deleteById
      description: "Deleta registro de vacina pelo id"
      responses:
        200:
          description: "ok"
          schema:
            type: string
        400:
          description: "Consulta inválida"
        401:
          description: "Não foi possível credenciais para autenticação inválida"
        404:
          description: "Vacina não encontrado"
        500:
          description: "Erro interno"
      parameters:
        - in: "path"
          name: "id"
          type: "integer"
          format: int64
          required: true
          description: "id da vacina"


definitions:
  VacinaDTO:
    type: object
    properties:
      id:
        type: integer
        format: int64
        description: "Identificador da vacina"
      nome:
        type: "string"
        description: "Nome da vacina"
      marca:
        type: "string"
        description: "Marca da vacina"
      lote:
        type: "string"
        description: "Lote da vacina"
  VacinaRequest:
    type: object
    required:
      - nome
      - marca
      - lote
    properties:
      nome:
        type: "string"
        description: "Nome da vacina"
      marca:
        type: "string"
        description: "Marca da vacina"
      lote:
        type: "string"
        description: "Lote da vacina"
        
        
        
        
        
        
        
        
        
        
        
        
        
         
      